+++ app/models/grouping_starter_file_entry.rb
@@ -3,4 +3,8 @@
 class GroupingStarterFileEntry < ApplicationRecord
   belongs_to :starter_file_entry
   belongs_to :grouping
+
+  validates_presence_of :grouping
+  validates_presence_of :starter_file_entry
+  validates_uniqueness_of :starter_file_entry_id, scope: :grouping_id
 end