+++ spec/factories/section_starter_file_groups.rb
@@ -1,0 +1,6 @@
+FactoryBot.define do
+  factory :section_starter_file_group do
+    association :section
+    association :starter_file_group
+  end
+end