+++ spec/models/assignment_properties_spec.rb
@@ -3,6 +3,7 @@ describe AssignmentProperties do
     it { is_expected.to validate_presence_of(:repository_folder) }
     it { is_expected.to validate_presence_of(:group_min) }
     it { is_expected.to validate_presence_of(:group_max) }
+    it { is_expected.to validate_inclusion_of(:starter_file_type).in_array(AssignmentProperties::STARTER_FILE_TYPES) }
 
     it { is_expected.to validate_numericality_of(:group_min).is_greater_than(0) }
     it { is_expected.to validate_numericality_of(:group_max).is_greater_than(0) }